requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
Brotli==1.1.0
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
contextlib2==21.6.0
crewai==0.28.8
crewai-tools==0.1.4
cryptography==42.0.5
dataclasses-json==0.6.4
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
docstring-parser==0.15
embedchain==0.1.99
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-aiplatform==1.50.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.4.1
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gptcache==0.1.43
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
install==1.3.5
instructor==0.5.2
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
lancedb==0.5.7
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.50
langchain-experimental==0.0.57
langchain-google-genai==1.0.1
langchain-google-vertexai==1.0.3
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langsmith==0.1.38
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.14.3
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pluggy==1.4.0
posthog==3.5.0
proto-plus==1.23.0
protobuf==4.25.3
pulsar-client==3.4.0
py==1.11.0
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PyGithub==1.59.1
Pygments==2.17.2
PyJWT==2.8.0
pylance==0.9.18
PyNaCl==1.5.0
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.0.0
pyright==1.1.356
pysbd==0.3.4
PySocks==1.7.1
pysqlite-binary==0.5.1.3380300
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytube==15.0.0
PyYAML==6.0.1
ratelimiter==1.2.0.post0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.1
rsa==4.9
schema==0.7.5
selenium==4.19.0
semver==3.0.2
shapely==2.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.2
tomli==2.0.1
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.4
types-protobuf==4.25.0.20240417
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
zipp==3.18.1
